chore: extract OpenQueryButton to a separate container [YTFRONT-4463]
